pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^5.0.2"
django-bootstrap5 = "^23.4"
django-extensions = "^3.2.3"
ipython = "^8.21.0"
django-filter = "^24.1"
[tool.poetry.dev-dependencies]
flake8 = "^7.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
